+++ backend/canisters/storage_index/impl/src/model/mod.rs
@@ -1,3 +1,4 @@
+pub mod bucket_event_batch;
 pub mod bucket_sync_state;
 pub mod buckets;
 pub mod files;